New features, new models.
- Added subscription screen.
- Added consumables screen.
- UI refactoring.
- Component building.

+++ SuperPage/Features/Usage/Model/Tokens.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct Tokens: Codable {
     
-    var purchased: Int?
+    var purchased: Double?
 }
